Revert "chore: Code style standardization and root directory cleanup"
This reverts commit a275db3fdc51131492ab92a35e863c83557b8776.

+++ src/GauntletCI.Tests/DiffParserExtendedTests.cs
@@ -40,10 +40,7 @@ public class DiffParserExtendedTests
         }
         finally
         {
-            if (File.Exists(tempFile))
-            {
-                File.Delete(tempFile);
-            }
+            if (File.Exists(tempFile)) File.Delete(tempFile);
         }
     }
